Initial commit
This creates a Swift package with an example app, linting and formatting
commands, CI pipeline, README, and contributing guidelines.

In the absence of guidance or data about which minimum Swift version to
use, I decided to use the latest release at time of writing (5.10). We
can revisit this later if need be.

The SwiftLint rules are taken from our Asset Tracking repo at commit
b39e5187 (plus one tweak to a comment). They seemed like a good starting
point (I had previously done some thinking about these rules in [1]),
and we can tweak them as needed. I’ve enabled the private_swiftui_state
rule (seems useful) and also inverted its default trailing comma rule
(for compatibility with SwiftFormat).

Upon realising that SwiftLint and SwiftFormat have clashing opinions on
trailing commas, I worried that there might turn out to be other clashes
between the two. So I then hoped that we could get away with just using
SwiftLint, which does have some style rules. But it doesn’t enforce some
basic things like indentation. I hope that there aren’t further clashes
between the two tools, but we won’t find out until we write more code.

We use Mint to run SwiftLint and SwiftFormat. This allows us to lock
their versions, which we can’t do with Homebrew. It’s unfortunate that
for some reason SwiftLint takes an eternity to build using Mint, but
it’s a one-time thing per developer. I originally ran these tools as
Swift package plugins, but SwiftLint’s plugin was crashing when run in
Xcode 16 beta, hence the switch to Mint. Maybe we can switch back
sometime.

The build scripts used by CI, as well as the linting scripts, are
implemented as a Swift command-line tool. Hopefully we’ll find this
easier to maintain and modify than a set of Bash scripts. Note that in
these scripts I’m using APIs that are not available in macOS 11, which
is the package’s deployment target. I’m happy for this tool to only work
on macOS 14 (i.e. the latest at time of writing) but there doesn’t seem
to be a way to express per-target platform requirements in SPM. So
instead I’ve had to pepper the code with #available(macOS 14, *)
annotations; don’t know if there’s a better way to do it (I tried
creating a separate, nested Package.swift but Xcode refused to work with
it).

The library’s minimum supported OS versions (i.e. iOS 14 and
equivalents) are taken from Umair’s guidance [2] which I believe relates
to the accepted RFC for increasing ably-cocoa’s minimum-supported OS
versions to the same [3]. It’s not yet clear to me how we’ll guarantee
support for these versions given that, at the time of writing, the
oldest iOS simulator that you could install in the latest Xcode version
(15.4) is 15.0. For now, I’m just testing on the latest release OS
versions at time of writing.

As for which Xcode versions we test on in CI, I’m (if you ignore the
beta Xcode 16 testing) currently just testing on the oldest version of
Xcode that we support (15.3). We might want to expand this to test new
versions of Xcode as they’re released (because sometimes new compiler
warnings are introduced, I think).

I haven’t committed the `.swiftpm/configuration/Package.resolved` file
created by Xcode 16. We already have two Package.resolved files (see the
`comparePackageLockfiles` lint task) and so let’s wait for Xcode 16 to
be released to see whether this file is going to stick around (perhaps
it will become part of the .gitignore created by the SPM that ships with
Xcode 16).

[1] https://github.com/ably/ably-asset-tracking-swift/pull/629
[2] https://ably.atlassian.net/wiki/spaces/SDK/pages/3261202438/Swift+Architecture+Thoughts
[3] https://ably.atlassian.net/wiki/spaces/ENG/pages/3199500291/ADR-114+Increase+Cocoa+SDK+minimum+supported+version+to+iOS+14