Implement room status change upon attach or detach
Based on the simplified requirements described in #19.

The decision from 7d6acde to use actors as the mechanism for managing
mutable state means that I’ve had to make RoomStatus.{ current, error,
onChange(bufferingPolicy:) } async. As mentioned there, if later on we
decide this is too weird an API, then we can think of alternatives.

I really wanted to avoid making DefaultRoomStatus an actor; I tried to
find a way to isolate this state to the DefaultRoom actor (who logically
owns this state), by trying to make the DefaultRoomStatus access the
DefaultRoom-managed state, but I was not successful and didn’t want to
sink much time into it. It means that DefaultRoom has suspension points
in order to access its current state, which I am not happy about.  But
we can revisit later, perhaps armed with more knowledge of concurrency
and in less of a rush to get some initial functionality implemented.

Resolves #19.