Implement logging
The aim of this commit is to:

- provide infrastructure for internal components to be able to log
  messages
- add some log messages to the existing functionality (I am not
  particularly concerned by the quality or utility of these log messages
  at the moment)
- provide a basic default logging backend so that users can see log
  messages in the console

The choice to have a separate internal logging interface is taken from
Asset Tracking (AAT) and ably-cocoa, where I think it works well. I
haven’t currently copied across AAT’s subsystem mechanism, a large part
of the motivation for which was to provide a way to integrate the logs
emitted by the AAT-instantiated Realtime client into those emitted by
AAT itself; this is not currently a goal for Chat, which does not
instantiate a Realtime client.

I briefly toyed with the idea of writing a macro that would log the
inputs and outputs of public API methods, but figuring out how to do
that didn’t seem like a good use of time now.

Resolves #38.