Only support Swift 6 in example app
We want to unlock Swift 6 features in the example app, the same way that
67f4563 did so for the SDK.

It might be worth having some other small project that verifies that the
SDK can be used by a project that targets Swift 5, but I don’t think
this should be the example app, whose code should look modern and shiny.
Created #57 to look into this sometime.