Add a convention for exposing APIs to tests
I had noticed that I was writing lots of comments like "Exposed for
testing" and thought it would be useful to do something more consistent
and lintable.

I’ve gated the code for re-exposing APIs behind a #if DEBUG, to avoid
bloating release builds. Now that there’s a divergence between the Debug
and Release configuration, we should update CI to also build in the
Release configuration; have created #68 for this.